test get all do controller passando

+++ ws-banco-tabajara/Ws-BancoTabajara/Ws-BancoTabajara.Api/Controllers/Clients/ClientsController.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using FluentValidation;
+using FluentValidation.Results;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
@@ -7,6 +9,9 @@ using System.Web.Http;
 using Ws_BancoTabajara.Api.Controllers.Common;
 using Ws_BancoTabajara.Api.Extensions;
 using Ws_BancoTabajara.Applications.Features.Clients;
+using Ws_BancoTabajara.Applications.Features.Clients.Commands;
+using Ws_BancoTabajara.Applications.Features.Clients.Queries;
+using Ws_BancoTabajara.Applications.Features.Clients.ViewModels;
 using Ws_BancoTabajara.Domain.Features.Clients;
 using Ws_BancoTabajara.Infra.ORM.Base;
 using Ws_BancoTabajara.Infra.ORM.Features.Clients;
@@ -18,18 +23,17 @@ namespace Ws_BancoTabajara.Api.Controllers.Clients
     {
         public IClientService _clientService;
 
-        public ClientsController()
+        public ClientsController(IClientService clientService) : base()
         {
-            var context = new BancoTabajaraDbContext();
-            var repository = new ClientRepository(context);
-            _clientService = new ClientService(repository);
+            _clientService = clientService;
         }
 
         [HttpGet]
         public IHttpActionResult GetAll()
         {
             int quantity = Request.GetQueryQuantityValueExtension();
-            return HandleQueryable<Client>(_clientService.GetAll(quantity));
+            var query = new ClientQuery { Quantity = quantity };
+            return HandleQueryable1<Client, ClientViewModel>(_clientService.GetAll(query));
         }
 
         [HttpGet]
@@ -40,20 +44,32 @@ namespace Ws_BancoTabajara.Api.Controllers.Clients
         }
 
         [HttpPost]
-        public IHttpActionResult Add(Client client)
+        public IHttpActionResult Add(ClientRegisterCommand client)
         {
+            var validate = client.Validate();
+            if (!validate.IsValid)
+                return HandleValidationFailure(validate.Errors);
+
             return HandleCallback(() => _clientService.Add(client));
         }
 
         [HttpPut]
-        public IHttpActionResult Update(Client client)
+        public IHttpActionResult Update(ClientUpdateCommand client)
         {
+            var validate = client.Validate();
+            if (!validate.IsValid)
+                return HandleValidationFailure(validate.Errors);
+
             return HandleCallback(() => _clientService.Update(client));
         }
 
         [HttpDelete]
-        public IHttpActionResult Remove(Client client)
+        public IHttpActionResult Remove(ClientRemoveCommand client)
         {
+            var validate = client.Validate();
+            if (!validate.IsValid)
+                return HandleValidationFailure(validate.Errors);
+
             return HandleCallback(() => _clientService.Remove(client));
         }
     }
